+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+# Head
+
+- Chore: used the `webpack` watch system for watching changes.
+
 # 0.0.1-alpha.5 - 2016-12-11
 
 - Chore: used the `make` event for the compilation fonts and styles.